+++ subsys/logging/CMakeLists.txt
@@ -5,6 +5,7 @@ if(NOT CONFIG_LOG_MINIMAL)
     CONFIG_LOG
     log_list.c
     log_core.c
+    log_mgmt.c
     log_msg.c
     log_output.c
   )
